+++ containers/schedule/ScheduleSummary/index.tsx
@@ -9,7 +9,6 @@ import SVG_share from "@/public/svgs/share.svg";
 import SVG_chart from "@/public/svgs/barchart.svg";
 import Button from "@/components/Button";
 import type { ScheduleSummaryType } from "@/types/schedule";
-import axios from "axios";
 
 import { useRecoilValue } from "recoil";
 import { selectRecoilDays } from "@/states/Schedule/atom";
