--- hooks/schedule/usePostSchedule.ts
@@ -1,3 +1,0 @@
-import axiosInstance from "@/components/Axiosinstance";
-import { queryKeys } from "@/components/Providers/querykey";
-import { useQueryClient, useMutation } from "@tanstack/react-query";
